11 - Desenvolvimento de APIs com Flask/dio_bank/pyproject.toml
[project]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = [
    {name = "Dawidson Pereira",email = "dawidsongru@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-migrate",
    "flask-jwt-extended"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
